requirement.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1

langchain==0.3.15
langchain-community==0.3.15
langchain-openai==0.2.5

pinecone-client==2.2.4

pypdf==4.2.0
aiofiles
requests
python-multipart
sentence-transformers==2.7.0
langchain-pinecone
